+++ example/type_zoo/type_zoo.ml
@@ -1,6 +1,5 @@
 open Base
 module Exposition = Type_zoo_pc.Exposition
-module Platypus = Type_zoo_pc.Platypus
 
 let message =
   Exposition.
@@ -12,9 +11,9 @@ let message =
       elephant = 46;
       fox = 47;
       giraffe = 48;
-      hest = [49; 50; 51];
-      indri = 52;
-      jellyfish = 53;
+      hest = [49L; 50L; 51L];
+      indri = 52l;
+      jellyfish = 53L;
       kingfisher = 54.1;
       llama = 55.2;
       meerkat = true;
@@ -32,9 +31,9 @@ let message =
             elephant = 0;
             fox = 0;
             giraffe = 0;
-            hest = [0; 0; 0];
-            indri = -52;
-            jellyfish = -53;
+            hest = [0L; 0L; 0L];
+            indri = -52l;
+            jellyfish = -53L;
             kingfisher = -54.1;
             llama = -55.2;
             meerkat = false;
@@ -53,6 +52,6 @@ let () =
   |> Exposition.to_binary
   >>= Exposition.of_binary
   >>| Exposition.show
-  >>| Printf.sprintf "Here's the deserialized ZOO:\n%s"
+  >>| Printf.sprintf "Here's the deserialized ZOO:\n\n%s"
   >>| Stdio.print_endline
   |> ignore